Отключает hoc withactiveplayer
