+++ src/components/tabs/tabs.jsx
@@ -1,13 +1,47 @@
-import React from "react";
-import PropTypes from "prop-types";
+import React, {useState} from "react";
+import FilmOverview from '../../components/film-overview/film-overview';
+import FilmDetails from '../../components/film-details/film-details';
+import FilmReviews from '../../components/film-reviews/film-reviews';
+import {validComments, validFilm} from "../../utils/props";
+
+const Tab = {
+  OVERVIEW: `OVERVIEW`,
+  DETAILS: `DETAILS`,
+  REVIEWS: `REVIEWS`,
+};
 
 const Tabs = (props) => {
-  const {onSwitchTab, tabToRender, poster, activeTab, tabs} = props;
+  const {film, comments} = props;
+  const tabs = Object.values(Tab);
+
+  const [activeTab, setState] = useState(Tab.OVERVIEW);
+
+  const handleSwitchTab = (tabName) => {
+    setState(tabName);
+  };
+
+  const renderTab = () => {
+    switch (activeTab) {
+      case Tab.OVERVIEW:
+        return <FilmOverview
+          film={film}
+        />;
+      case Tab.DETAILS:
+        return <FilmDetails
+          film={film}
+        />;
+      case Tab.REVIEWS:
+        return <FilmReviews
+          comments={comments}
+        />;
+    }
+    return null;
+  };
 
   return (
     <div className="movie-card__info">
       <div className="movie-card__poster movie-card__poster--big">
-        <img src={poster} alt="The Grand Budapest Hotel poster" width="218" height="327" />
+        <img src={film.poster} alt="The Grand Budapest Hotel poster" width="218" height="327" />
       </div>
       <div className="movie-card__desc">
         <nav className="movie-nav movie-card__nav">
@@ -19,7 +53,7 @@ const Tabs = (props) => {
                   className={activeTab === item ? `movie-nav__item movie-nav__item--active` : `movie-nav__item` }
                   onClick={(evt) => {
                     evt.preventDefault();
-                    onSwitchTab(item);
+                    handleSwitchTab(item);
                   }}
                 >
                   <a href="#" className="movie-nav__link">{item.charAt(0) + item.slice(1).toLowerCase()}</a>
@@ -28,18 +62,15 @@ const Tabs = (props) => {
             })}
           </ul>
         </nav>
-        {tabToRender}
+        {renderTab()}
       </div>
     </div>
   );
 };
 
 Tabs.propTypes = {
-  activeTab: PropTypes.string.isRequired,
-  onSwitchTab: PropTypes.func.isRequired,
-  tabToRender: PropTypes.element.isRequired,
-  poster: PropTypes.string.isRequired,
-  tabs: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
+  film: validFilm,
+  comments: validComments,
 };
 
 export default Tabs;
